pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


requires-python = ">=3.12"

[tool.poetry]
name = "discord-blue"
version = "0.1.0"
description = ""
authors = ["Chris Busillo <info@shinycomputers.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
toml = "^0.10.2"
nextcord = "^2.5.1"
printnodeapi = "^0.2.0"
pynacl = "^1.5.0"
modern-shippo = "3.0.1"
reportlab = "^4.2.4"
svglib = "^1.5.1"
discord-py = { extras = ["voice"], version = "^2.4.0" }
aiofiles = "^24.1.0"
pydantic = "^2.9.2"
torch = "^2.4.1"
transformers = "^4.45.1"
accelerate = "^0.34.2"
datasets = "^3.0.1"
sentencepiece = "^0.2.0"

[tool.poetry.group.dev.dependencies]
setuptools = "^72.1.0"
black = "^24.8.0"
types-toml = "^0.10.8.7"

[tool.poetry.scripts]
discord-blue = 'discord_blue.__main__:main'

[tool.mypy]
strict = true
python_version = 3.12
warn_any_return = false
warn_no_untyped_call = false
